Ajna max generate and max withdraw calcs adjustment
