+++ packages/dma-library/src/protocols/ajna/consts.ts
@@ -1,0 +1,1 @@
+export const ajnaCollateralizationFactor = 1.04
